Add `--stash` as alternative to `--clean` to `update-checkout`
Use of `--clean` can lead to irreversible loss of uncommitted data. It is still desirable to reset Swift project repositories to a clean state, but without deleting all of the in-progress changes. Passing `--stash` instead of (or in addition to) `--clean` will preserve uncommitted changes in stashes of corresponding repositories instead of completely deleting those.