[FSPrunedLiveness] Returned bitfield from isDef.
Whether a node is a def on a collection of bits (whether a range or a
bit vector) isn't exhaustively characterized by the values {true,
false}.  A node may be a def on some but not others of the bits in the
collection.  Changed the range-taking isDef to write back the bits in
the range at which the node is a def.  The caller can then decide how to
react with full information.