[MoveChecker] Don't add defs to liveness.
When extending liveness, the instructions prior to a preexisting destroy
are added to liveness.  If that prior instruction is the def, adding it
to liveness results in multi-def liveness understanding that there is a
prior def.

Fixes a bug exposed by adding coroutine ends to liveness rather than
function ends.