Expose Split-DWARF support in Frontend and Driver
This patch wires up flags in Driver and Frontend that expose support
for GNU debug fission / DWARF5 standardized Split-DWARF as implemented
in LLVM. It splits the majority of debug information from .o outputs
into separate .dwo files.

This is based on #70296 authored by weliveindetail and adds a rebase,
fix for the macos test failure and some flag renaming.