Introduce an assertion configuration for "release with bounds safety"
To support the introduction of bounds checks for the unsafe buffer pointer
types, introduce a new assertion confirmation that is for release builds
that also have this buffer safety. It can be enabled with the experimental
feature `UnsafePointerBoundsSafety` in normal builds, or (for testing)
the `ReleaseWithBoundsSafety` assert configuration.

The end goal of this approach to change the experimental feature to
an upcoming feature, which will also enabled with Swift 6.