[Test] Make the ExternalGenericMetadataBuilder test more lenient with equality.
Aspects of the emitted JSON vary depending on the exact contents of the built binaries. This happens when we refer to a private symbol with an offset from a public symbol. The offset and choice of public symbol may vary. Filter these out of the JSON before checking for equality.