Sema: test @_spiOnly with warnings about superfluously public imports
The logic reporting superfluously public imports treats @_spiOnly
imports as any public import. As such @_spiOnly shouldn't interfere with
the diagnosics. Let's make sure we test the current behavior.

In the future, we may want to extend the existing logic to report bare
`public imports` that could be `@_spiOnly public import` if referred
from SPI but not API.