Sema: track use of public imports by more extensions
Consider extensions to imported types as users of a public import. This
silences false reports of unused public imports in the case of an empty
extension satisfying generic requirements via retroactive conformances.
It also silences real reports of superfluously public imports used in
internal extension, but this isn't very harmful at this time.

We should revisit this later and tighten the check, we should refactor
the logic use to determine what makes an extension public.

rdar://120391841