Sema: Fix recent regression with -disable-experimental-associated-type-inference
Change the order of bindings to exactly match what was going on before,
because we seem to hit a request cycle via AssociatedConformanceRequest.

Of course the latter is lazier with -enable-experimental-associated-type-inference,
and the staging flag will be ripped out completely soon, but we want to keep the
old mode working a little while longer.

Fixes rdar://problem/122810266.