Builtins: do not allowInverses
When building a generic signature for a builtin function, do not infer
Copyable and Escapable requirements. We should be explicit about what
invertible protocols are required of the generic parameters for a
builtin. This will require an audit of our existing Builtins to add the
requirements as-needed, which is a task I'm deferring for now.