[Concurrency] Don't skip conformance derivation for types with isolated stored properties.
Otherwise, types that rely on derived conformances to Equatable and friends will
fail to compile if a library adds a `@preconcurrency @MainActor` annotation. There
is no need to prevent conformance derivation, because the actor isolation checker
still runs on derived witnesses and will diagnose any concurrecy violations (with
appropriate downgrading behavior for `@preconcurrency`).